Removed the need for the Peewee ORM and replaced with builtin sqlite3. Moved code into a sinlge file.
